Authenticate and Authoraization

+++ src/tasks/tasks.service.ts
@@ -1,5 +1,7 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { query } from 'express';
+import { User } from 'src/auth/user.entity';
 // import { Task, TaskStatus } from './tasks.mode';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { GetTaskFilterDto } from './dto/get-task-filter.dto';
@@ -14,20 +16,22 @@ export class TasksService {
             private taskRepository : TaskRepository
         ){}
 
-        async getTasks(filterDto: GetTaskFilterDto):Promise<Task[]>{
-          return this.taskRepository.getTasks(filterDto)
+        async getTasks(filterDto: GetTaskFilterDto, user:User):Promise<Task[]>{
+          return this.taskRepository.getTasks(filterDto,user)
         
         }
-        async getTaskById(id:number): Promise<Task>{
-            const found = await this.taskRepository.findOne(id);
+        async getTaskById(id:number,user: User): Promise<Task>{
+            const found = await this.taskRepository.findOne({where :{id,userId:user.id}});
             if (!found){
                             throw new NotFoundException(`Taks with ID ${id} not found`)
                         }
             return found;            
         }
 
-        async deleteTask(id:number):Promise<void>{
-          const result = await this.taskRepository.delete(id);
+        async deleteTask(id:number,
+          user:User
+          ):Promise<void>{
+          const result = await this.taskRepository.delete({id,userId: user.id});
           
           if(result.affected === 0){
             throw new NotFoundException(`Tasks with ID ${id} not found`)
@@ -35,8 +39,10 @@ export class TasksService {
         }
 
 
-        async createTask(createTaskDto: CreateTaskDto):Promise<Task>{
-         return this.taskRepository.createTask(createTaskDto);
+        async createTask(createTaskDto: CreateTaskDto,
+          user:User
+          ):Promise<Task>{
+         return this.taskRepository.createTask(createTaskDto,user);
         }
 //     getTaskWithFilter(filterDto:GetTaskFilterDto):Task[]{
 //        const {status, search} = filterDto;
@@ -52,16 +58,17 @@ export class TasksService {
 //        return tasks;
 //     }
       
-        async updateTaskStatus(id : number, status: TaskStatus): Promise<Task>{
-          const task = await this.getTaskById(id);
+        async updateTaskStatus(id : number, status: TaskStatus, user: User): Promise<Task>{
+          const task = await this.getTaskById(id,user);
           task.status= status;
           await task.save();
           return task;
         }
-//     updateTaskStatus(id:string, status: TaskStatus):Task{
-//         const task = this.getTaskById(id);
-//         task.status = status;
-//         return task;
-//     }
+    // async updateTaskStatus(id:number, status: TaskStatus,user:User):Task{
+    //     const task = this.getTaskById(id,user);
+    //     task.status = status;
+    //     await task.save()
+    //     return task;
+    // }
 }
